Fix `-Wsign-compare` issues
Cast variables from larger types to smaller types if the range/bounds is
sufficiently small, when doing comparisons between the two values.

Apply stylistic fixes/suggestions from checkpatch.pl as need be.

+++ testcases/open_posix_testsuite/conformance/interfaces/pthread_cond_destroy/2-1.c
@@ -437,7 +437,7 @@ int main(void)
 	}
 
 	/* Do the test for each test scenario */
-	for (scenar = 0; scenar < NSCENAR; scenar++) {
+	for (scenar = 0; scenar < (int)NSCENAR; scenar++) {
 		/* set / reset everything */
 		td->fork = 0;
 		ret = pthread_mutexattr_init(&ma);
