+++ testcases/open_posix_testsuite/conformance/interfaces/pthread_mutex_trylock/4-2.c
@@ -256,7 +256,7 @@ int main(void)
  * For each test scenario, initialize the attributes and other variables.
  * Do the whole thing for each time to test.
  */
-	for (sc = 0; sc < NSCENAR; sc++) {
+	for (sc = 0; sc < (int)NSCENAR; sc++) {
 #if VERBOSE > 1
 		output("[parent] Preparing attributes for: %s\n",
 		       scenarii[sc].descr);
