+++ testcases/open_posix_testsuite/conformance/interfaces/pthread_sigmask/7-1.c
@@ -41,7 +41,7 @@ void *a_thread_func()
 		SIGTRAP, SIGURG, SIGVTALRM, SIGXCPU, SIGXFSZ
 	};
 
-	for (i = 0; i < NUMSIGNALS; i++) {
+	for (i = 0; i < (int)NUMSIGNALS; i++) {
 		sigemptyset(&oactl);
 		sigemptyset(&tempset);
 		sigaddset(&tempset, siglist[i]);
